Split all un/binop tests in one file per operator
The previous code organization meant that path condition from the first ifs were carried out to the following ifs, so the operators tested at the bottom of the file were not tested on all the possible input values. By splitting each operator's test in its own file, we ensure that the each operator is tested on all possible inputs.